+++ ports/efm8/efm8/efm8_util/reset.h
@@ -8,20 +8,20 @@
 /// @brief  Perform a software reset
 /// @note   This function is implemented as a macro.
 #ifdef IS_DOXYGEN
-extern void efm8_software_reset(void);
+    extern void efm8_software_reset(void);
 #else
     #if defined(EFM8UB3_DEVICE) || defined(EFM8UB1_DEVICE)
         // reset on page 0
         #define efm8_software_reset() \
             do { \
                 SFRPAGE = 0x00; \
-                RSTSRC  = RSTSRC_SWRSF__BMASK; \
+                RSTSRC  = RSTSRC_SWRSF__BMASK | RSTSRC_PORSF__BMASK; \
             } while (0)
     #elif defined(EFM8UB2_DEVICE)
         // reset on all pages
         #define efm8_software_reset() \
             do { \
-                RSTSRC  = RSTSRC_SWRSF__BMASK; \
+                RSTSRC  = RSTSRC_SWRSF__BMASK | RSTSRC_PORSF__BMASK; \
             } while (0)
     #endif
 #endif
@@ -30,7 +30,7 @@ extern void efm8_software_reset(void);
 /// @brief  Perform a software reset, and enter the factor bootloader
 /// @note   This function is implemented as a macro.
 #ifdef IS_DOXYGEN
-extern void efm8_factory_bootloader_jump(void);
+    extern void efm8_factory_bootloader_jump(void);
 #else
     #if defined(EFM8UB3_DEVICE) || defined(EFM8UB1_DEVICE)
         // reset on page 0
@@ -38,14 +38,14 @@ extern void efm8_factory_bootloader_jump(void);
             do { \
                 SFRPAGE = 0x00; \
                 *((uint8_t __idata*) 0x00) = 0xA5; \
-                RSTSRC = RSTSRC_SWRSF__BMASK; \
+                RSTSRC = RSTSRC_SWRSF__BMASK | RSTSRC_PORSF__BMASK; \
             } while (0)
     #elif defined(EFM8UB2_DEVICE)
         // reset on all pages
         #define efm8_factory_bootloader_jump() \
             do { \
                 *((uint8_t __idata*) 0x00) = 0xA5; \
-                RSTSRC = RSTSRC_SWRSF__BMASK; \
+                RSTSRC = RSTSRC_SWRSF__BMASK | RSTSRC_PORSF__BMASK; \
             } while (0)
     #endif
 #endif